requirements.txt
Flask
request
render_template
redirect
url_for
werkzeug.utils
secure_filename
torchvision
torchvision.transforms
os
CNNNet
torch
matplotlib.pyplot
PIL
Image
